Add files via upload
The main aim of this project is to simulate the working of Dijkstra’s algorithm by making use 
of OpenGL computer graphics. The computer graphics help us to understand better the working 
of Dijkstra’s algorithm as it is visualized. Main aim of this project is to implement and 
understand the basic OpenGL functions and accurately know the working of it, which is an 
important aspect of computer graphics.